requirements.txt
tensorflow
requests
numpy
pefile
fastapi
uvicorn
py-cpuinfo
psutil
nvidia-ml-py3
python-magic
vt-py
msgpack
capstone
python-multipart